+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - When an execution contains a failed step the exit code will now be `1`
 
+### BUG FIXES
+
+- Fix when using "just in time" tfvars from param store 
+
 ## 0.0.9 (Apr 20, 2020)
 
 ### ENHANCEMENTS